app: Reintroduce avatar input for attributes
