+++ app/src/components/form/attribute_input.rs
@@ -4,6 +4,8 @@ use yew::{
     function_component, html, use_effect_with_deps, use_mut_ref, use_node_ref, virtual_dom::AttrValue, Component, Context, Html, Properties
 };
 
+use super::file_input::JpegFileInput;
+
 /*
  <input
                 ref={&ctx.props().input_ref}
@@ -28,7 +30,11 @@ fn attribute_input(props: &AttributeInputProps) -> Html {
         AttributeType::String => "text",
         AttributeType::Integer => "number",
         AttributeType::DateTime => "datetime-local",
-        AttributeType::Jpeg => "file",
+        AttributeType::Jpeg => {
+            return html! {
+                <JpegFileInput name={props.name.clone()} value={props.value.clone()} />
+            }
+        },
     };
     let accept = match props.attribute_type {
         AttributeType::Jpeg => Some("image/jpeg"),
