+++ app/src/components/form/mod.rs
@@ -1,6 +1,7 @@
 pub mod attribute_input;
 pub mod checkbox;
 pub mod field;
+pub mod file_input;
 pub mod select;
 pub mod static_value;
 pub mod submit;
